Moved resource interfaces into the new Interfaces.Resource namespace.
