+++ SS14.Server/BaseServer.cs
@@ -38,6 +38,7 @@ using SS14.Shared.Reflection;
 using SS14.Shared.Timing;
 using SS14.Shared.Utility;
 using SS14.Shared.Interfaces.Log;
+using SS14.Shared.Interfaces.Resources;
 
 namespace SS14.Server
 {
