+++ SS14.Shared/ContentPack/AssemblyLoader.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using SS14.Shared.Interfaces;
 using SS14.Shared.Interfaces.Reflection;
+using SS14.Shared.Interfaces.Resources;
 using SS14.Shared.IoC;
 using SS14.Shared.Log;
