+++ SS14.Shared/ContentPack/ResourceManager.cs
@@ -4,6 +4,7 @@ using System.IO;
 using SS14.Shared.Configuration;
 using SS14.Shared.Interfaces;
 using SS14.Shared.Interfaces.Configuration;
+using SS14.Shared.Interfaces.Resources;
 using SS14.Shared.IoC;
 using SS14.Shared.Log;
 using SS14.Shared.Utility;
