+++ SS14.Shared/SS14.Shared.csproj
@@ -144,7 +144,6 @@
     <Compile Include="ContentPack\GameServer.cs" />
     <Compile Include="ContentPack\GameShared.cs" />
     <Compile Include="ContentPack\IContentRoot.cs" />
-    <Compile Include="ContentPack\IWritableDirProvider.cs" />
     <Compile Include="ContentPack\PackLoader.cs" />
     <Compile Include="ContentPack\PathHelpers.cs" />
     <Compile Include="ContentPack\ResourceManager.cs" />
@@ -181,7 +180,6 @@
     <Compile Include="Interfaces\GameObjects\IEntityQuery.cs" />
     <Compile Include="Interfaces\GameObjects\System\IEntitySystem.cs" />
     <Compile Include="Interfaces\GameObjects\IEntitySystemManager.cs" />
-    <Compile Include="Interfaces\IResourceManager.cs" />
     <Compile Include="Interfaces\Log\ILogHandler.cs" />
     <Compile Include="Interfaces\Log\ILogManager.cs" />
     <Compile Include="Interfaces\Log\ISawmill.cs" />
@@ -194,6 +192,8 @@
     <Compile Include="Interfaces\Physics\ICollider.cs" />
     <Compile Include="Interfaces\Physics\ICollisionManager.cs" />
     <Compile Include="Interfaces\Reflection\IReflectionManager.cs" />
+    <Compile Include="Interfaces\Resources\IResourceManager.cs" />
+    <Compile Include="Interfaces\Resources\IWritableDirProvider.cs" />
     <Compile Include="Interfaces\Serialization\ISS14Serializer.cs" />
     <Compile Include="Interfaces\Map\IMapChunk.cs" />
     <Compile Include="Interfaces\Map\IMapGrid.cs" />
